fixed edit delete actions only by creator

+++ SamsAuctions/BL/Auctions.cs
@@ -4,6 +4,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Security.Claims;
 using System.Threading.Tasks;
+using Microsoft.AspNetCore.Identity;
 using SamsAuctions.DAL;
 using SamsAuctions.Models;
 
@@ -12,18 +13,27 @@ namespace SamsAuctions.BL
     public class Auctions : IAuctions
     {
         IAuctionsRepository _repository;
+        private UserManager<AppUser> _userManager;
 
-        public Auctions(IAuctionsRepository repository)
+
+        public Auctions(IAuctionsRepository repository, UserManager<AppUser> userManager)
         {
             _repository = repository;
+            _userManager = userManager;
         }
+
 
         public async Task AddOrUpdateAuction(Auction auction, ClaimsPrincipal user)
         {
 
             if (user?.IsInRole("Admin") ?? false)
             {
-                await _repository.AddOrUpdateAuction(auction);
+                var currentUserName = (await _userManager.GetUserAsync(user)).UserName;
+
+                if (auction.AuktionID == 0 || (auction.AuktionID > 0 && currentUserName == auction.SkapadAv))
+                    await _repository.AddOrUpdateAuction(auction);
+                else
+                    throw new InvalidOperationException("User must have created this auction");
             }
             else
                 throw new InvalidOperationException("User must be in role: Admin");
@@ -44,7 +54,13 @@ namespace SamsAuctions.BL
             if (user?.IsInRole("Admin") ?? false)
             {
 
-                await _repository.RemoveAuction(auctionId, groupCode);
+                var currentUserName = (await _userManager.GetUserAsync(user)).UserName;
+                var auction = await GetAuction(auctionId, groupCode);
+
+                if (currentUserName == auction.SkapadAv)
+                    await _repository.RemoveAuction(auctionId, groupCode);
+                else
+                    throw new InvalidOperationException("User must have created this auction");
             }
             else
                 throw new InvalidOperationException("User must be in role: Admin");
